Put table output production in seperate module
Currently we have logic to produce tabular output, in main file. We
may add an option to show output in JSON so that stresster can be used
in automation pipelines. For that we need to add another logic and we
shall put that logic into seperate module too. So that clear seperation
can be achieved.