Put payload and headers in separate struct
Currently payload and headers are part of a payload type. Also there
are few other things like method, url which are shared separately. Share
everything separately is not optimal approach from the point of
maintainability as every time we introduce new entity to share, we
need to add separate parameter, create separate type which is tedious.
This commit starts grouping of payload and headers data for now. Eventually
that group will cover other shared entities too.