Get URL from data file and share it through shared data struct
Currently URL is received from command line arguments. This can be
a bit of inconvenient when URL is long. Also sometimes supplying it
as command line arg is bit of an issue. This commit receives URL from
a data file. It also changes how URL is shared by sharing it using Data
struct. Overall it does following changes.

1. Make data/paylaod file required
2. Make URL field required in data file
3. Share URL through shared Data struct instead of sharing separately